split enhy header from z64
