+++ include/functions.h
@@ -1478,21 +1478,7 @@ void SoundSource_UpdateAll(GlobalContext* globalCtx);
 void SoundSource_PlaySfxAtFixedWorldPos(GlobalContext* globalCtx, Vec3f* worldPos, u32 duration, u16 sfxId);
 void SoundSource_PlaySfxEachFrameAtFixedWorldPos(GlobalContext* globalCtx, Vec3f* worldPos, u32 duration, u16 sfxId);
 u16 ElfMessage_GetFirstCycleHint(GlobalContext* globalCtx);
-s32 EnHy_ChangeAnim(SkelAnime* skelAnime, s16 animIndex);
-struct EnDoor* EnHy_FindNearestDoor(Actor* actor, GlobalContext* globalCtx);
-void EnHy_ChangeObjectAndAnim(EnHy* enHy, GlobalContext* globalCtx, s16 animIndex);
-s32 EnHy_UpdateSkelAnime(EnHy* enHy, GlobalContext* globalCtx);
-void EnHy_Blink(EnHy* enHy, s32 arg1);
-s32 EnHy_Init(EnHy* enHy, GlobalContext* globalCtx, FlexSkeletonHeader* skeletonHeaderSeg, s16 animIndex);
-void func_800F0BB4(EnHy* enHy, GlobalContext* globalCtx, struct EnDoor* door, s16 arg3, s16 arg4);
-s32 func_800F0CE4(EnHy* enHy, GlobalContext* globalCtx, ActorFunc draw, s16 arg3, s16 arg4, f32 arg5);
-s32 func_800F0DD4(EnHy* enHy, GlobalContext* globalCtx, s16 arg2, s16 arg3);
-s32 EnHy_SetPointFowards(EnHy* enHy, GlobalContext* globalCtx, f32 gravity, s16 animIndex);
-s32 EnHy_SetPointBackwards(EnHy* enHy, GlobalContext* globalCtx, s16 animIndex);
-s32 EnHy_MoveForwards(EnHy* enHy, f32 speedTarget);
-s32 EnHy_MoveBackwards(EnHy* enHy, f32 speedTarget);
-void EnHy_UpdateCollider(EnHy* enHy, GlobalContext* globalCtx);
-s32 EnHy_PlayWalkingSound(EnHy* enHy, GlobalContext* globalCtx, f32 distAboveThreshold);
+
 u16 Text_GetFaceReaction(GlobalContext* globalCtx, u32 reactionSet);
 void EnvFlags_UnsetAll(GlobalContext* globalCtx);
 void EnvFlags_Set(GlobalContext* globalCtx, s16 flag);
