+++ include/variables.h
@@ -882,9 +882,6 @@ extern s801BB170 D_801BB170[118];
 extern Gfx D_801BC240[9];
 extern Gfx D_801BC288[3];
 // extern UNK_TYPE1 D_801BC2A0;
-extern s8 gEnHyBodyParts[];
-extern s8 gEnHyBodyPartsIndex[];
-extern u8 gEnHyShadowSize[];
 // extern UNK_TYPE1 D_801BC41E;
 extern ActorCutscene actorCutscenesGlobalCutscenes[8];
 extern s16 actorCutsceneCurrent;
